refactor: move routes to separate files
